+++ src/components/Movie.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 import Overdrive from 'react-overdrive'
 
-import { Movie as MovieInterface } from './MovieDetails'
+import { Movie as MovieInterface } from 'core/domains/Movie'
 
 const POSTER_PATH = 'https://image.tmdb.org/t/p/w154'
