+++ src/core/domains/Movie/index.ts
@@ -1,0 +1,8 @@
+export interface Movie {
+  id: string
+  backdrop_path: string
+  poster_path: string
+  title: string
+  release_date: string
+  overview: string
+}
